+++ internal/models/cards.go
@@ -12,7 +12,7 @@ type Card struct {
 	ListID      string       `json:"list_id"`
 	Title       string       `json:"title"`
 	Description string       `json:"description,omitempty"`
-	Position    int          `json:"position"`
+	Position    float64      `json:"position"`
 	DueDate     *time.Time   `json:"due_date,omitempty"`
 	Priority    CardPriority `json:"priority"`
 	Labels      []string     `json:"labels,omitempty"`
@@ -50,10 +50,10 @@ func NewCard(dbCard dbmodels.Card) Card {
 		d := dbCard.DeletedAt.Time
 		deleted = &d
 	}
-	pos := 0
+	pos := 0.0
 	if dbCard.Position.Big != nil {
 		f, _ := dbCard.Position.Big.Float64()
-		pos = int(f)
+		pos = f
 	}
 	return Card{
 		ID:          dbCard.ID,
